Only send command when LED state changes
Track if last sent command was on or off. When a new light level
message arrives, if the resulting command would be a repeat, then
we don't send it.

The reason for this is that the mosquitto test server actually
does queue all the messages up, and the Wio can send up updates
faster than it can receive them. So the nightlight started to lag
behind the light level, because it was wasting time handling
commands that didn't change things.

I believe the problem was brought to light because we reduced the
delay in the main loop on the Wio, from 5s to 2s. With the longer
delay, it shouldn't fall behind.

FIXME: The previous command variable is initialized to false/off,
so if the server is restarted, and during that time the light level
transitions from dark to bright, then the server will not submit
the off command until the level goes to dark first and then bright.
One fix would be to intialize to a string like "neither", which
will not be equal to either true or false, and so the first time
the command will always be sent.